+++ hammer.js
@@ -1,4 +1,4 @@
-/*! Hammer.JS - v1.1.0dev - 2014-04-08
+/*! Hammer.JS - v1.1.0dev - 2014-04-09
  * http://eightmedia.github.io/hammer.js
  *
  * Copyright (c) 2014 Jorik Tangelder <j.tangelder@gmail.com>;
@@ -103,12 +103,12 @@ Hammer.HAS_TOUCHEVENTS = ('ontouchstart' in window);
 
 
 /**
- * interval in which Hammer recalculates current velocity in ms
- * @property VELOCITY_INTERVAL
+ * interval in which Hammer recalculates current velocity/direction/angle in ms
+ * @property CALCULATE_INTERVAL
  * @type {Number}
  * @default 16
  */
-Hammer.VELOCITY_INTERVAL = 16;
+Hammer.CALCULATE_INTERVAL = 16;
 
 
 /**
@@ -409,6 +409,7 @@ var Utils = Hammer.utils = {
   getDirection: function getDirection(touch1, touch2) {
     var x = Math.abs(touch1.clientX - touch2.clientX)
       , y = Math.abs(touch1.clientY - touch2.clientY);
+
     if(x >= y) {
       return touch1.clientX - touch2.clientX > 0 ? DIRECTION_LEFT : DIRECTION_RIGHT;
     }
@@ -796,16 +797,13 @@ var Event = Hammer.event = {
       if(is_mouse && self.prevent_mouseevents) {
         return;
       }
-      // find out if we should detect a gesture.
-      // the touch/pointer/mouse must be at the screen or pressed.
-      // touchevents and pointerdown are always allowed
-      // and prevent mouseevents from being fired
-      else if(Utils.inStr(src_type, 'touch') || Utils.inStr(src_type, 'pointerdown')) {
-        self.prevent_mouseevents = true;
-        self.should_detect = true;
-      }
       // mousebutton must be down
       else if(is_mouse && ev.which === 1) {
+        self.should_detect = true;
+      }
+      // just a valid start event, but no mouse
+      else if(eventType == EVENT_START && !is_mouse) {
+        self.prevent_mouseevents = true;
         self.should_detect = true;
       }
 
@@ -831,7 +829,6 @@ var Event = Hammer.event = {
         PointerEvent.updatePointer(eventType, ev);
       }
     };
-
 
     this.bindDom(element, Hammer.EVENT_TYPES[eventType], bindDomOnTouch);
     return bindDomOnTouch;
@@ -1115,7 +1112,7 @@ var PointerEvent = Hammer.PointerEvent = {
 
 /**
  * @module hammer
- * 
+ *
  * @class Detection
  * @static
  */
@@ -1150,12 +1147,12 @@ var Detection = Hammer.detection = {
 
     // holds current session
     this.current = {
-      inst              : inst, // reference to HammerInstance we're working for
-      startEvent        : Utils.extend({}, eventData), // start eventData for distances, timing etc
-      lastEvent         : false, // last eventData
-      lastVelocityEvent : false, // last eventData for velocity.
-      velocity          : false, // current velocity
-      name              : '' // current gesture we're in/detected, can be 'tap', 'hold' etc
+      inst          : inst, // reference to HammerInstance we're working for
+      startEvent    : Utils.extend({}, eventData), // start eventData for distances, timing etc
+      lastEvent     : false, // last eventData
+      lastCalcEvent : false, // last eventData for calculations.
+      lastCalcData  : {}, // last lastCalcData
+      name          : '' // current gesture we're in/detected, can be 'tap', 'hold' etc
     };
 
     this.detect(eventData);
@@ -1226,62 +1223,40 @@ var Detection = Hammer.detection = {
 
 
   /**
-   * calculate velocity
+   * calculate velocity, angle and direction
    * @method getVelocityData
    * @param {Object} ev
    * @param {Number} delta_time
    * @param {Number} delta_x
    * @param {Number} delta_y
    */
-  getVelocityData: function getVelocityData(ev, delta_time, delta_x, delta_y) {
+  getCalculatedData: function getCalculatedData(ev, center, delta_time, delta_x, delta_y) {
     var cur = this.current
-      , velocityEv = cur.lastVelocityEvent
-      , velocity = cur.velocity;
+      , recalc = false
+      , calcEv = cur.lastCalcEvent
+      , calcData = cur.lastCalcData;
 
     // calculate velocity every x ms
-    if (velocityEv && ev.timeStamp - velocityEv.timeStamp > Hammer.VELOCITY_INTERVAL) {
-      velocity = Utils.getVelocity(ev.timeStamp - velocityEv.timeStamp,
-                                   ev.center.clientX - velocityEv.center.clientX,
-                                  ev.center.clientY - velocityEv.center.clientY);
-      cur.lastVelocityEvent = ev;
-    }
-    else if(!cur.velocity) {
-      velocity = Utils.getVelocity(delta_time, delta_x, delta_y);
-      cur.lastVelocityEvent = ev;
+    if (calcEv && ev.timeStamp - calcEv.timeStamp > Hammer.CALCULATE_INTERVAL) {
+      center = calcEv.center;
+      delta_time = ev.timeStamp - calcEv.timeStamp;
+      delta_x = ev.center.clientX - calcEv.center.clientX;
+      delta_y = ev.center.clientY - calcEv.center.clientY;
+      recalc = true;
     }
 
-    cur.velocity = velocity;
+    if(!cur.lastCalcEvent || recalc) {
+      calcData.velocity = Utils.getVelocity(delta_time, delta_x, delta_y);
+      calcData.angle = Utils.getAngle(center, ev.center);
+      calcData.direction = Utils.getDirection(center, ev.center);
 
-    ev.velocityX = velocity.x;
-    ev.velocityY = velocity.y;
-  },
-
-
-  /**
-   * calculate interim angle and direction
-   * @method getInterimData
-   * @param {Object} ev
-   */
-  getInterimData: function getInterimData(ev) {
-    var lastEvent = this.current.lastEvent
-      , angle
-      , direction;
-
-    // end events (e.g. dragend) don't have useful values for interimDirection & interimAngle
-    // because the previous event has exactly the same coordinates
-    // so for end events, take the previous values of interimDirection & interimAngle
-    // instead of recalculating them and getting a spurious '0'
-    if(ev.eventType == EVENT_END) {
-      angle = lastEvent && lastEvent.interimAngle;
-      direction = lastEvent && lastEvent.interimDirection;
-    }
-    else {
-      angle = lastEvent && Utils.getAngle(lastEvent.center, ev.center);
-      direction = lastEvent && Utils.getDirection(lastEvent.center, ev.center);
+      cur.lastCalcEvent = ev;
     }
 
-    ev.interimAngle = angle;
-    ev.interimDirection = direction;
+    ev.velocityX = calcData.velocity.x;
+    ev.velocityY = calcData.velocity.y;
+    ev.angle = calcData.angle;
+    ev.direction = calcData.direction;
   },
 
 
@@ -1293,14 +1268,11 @@ var Detection = Hammer.detection = {
    */
   extendEventData: function extendEventData(ev) {
     var cur = this.current
-      , startEv = cur.startEvent;
+      , startEv = cur.startEvent
+      , lastEv = cur.lastEvent || startEv;
 
-    // if the touches change, set the new touches over the startEvent touches
-    // this because touchevents don't have all the touches on touchstart, or the
-    // user must place his fingers at the EXACT same time on the screen, which is not realistic
-    // but, sometimes it happens that both fingers are touching at the EXACT same time
-    if(ev.touches.length != startEv.touches.length || ev.touches === startEv.touches) {
-      // extend 1 level deep to get the touchlist with the touch objects
+    // update the start touchlist to calculate the scale/rotation
+    if(ev.eventType == EVENT_TOUCH || ev.eventType == EVENT_RELEASE) {
       startEv.touches = [];
       Utils.each(ev.touches, function(touch) {
         startEv.touches.push(Utils.extend({}, touch));
@@ -1311,8 +1283,7 @@ var Detection = Hammer.detection = {
       , delta_x = ev.center.clientX - startEv.center.clientX
       , delta_y = ev.center.clientY - startEv.center.clientY;
 
-    this.getVelocityData(ev, delta_time, delta_x, delta_y);
-    this.getInterimData(ev);
+    this.getCalculatedData(ev, lastEv.center, delta_time, delta_x, delta_y);
 
     Utils.extend(ev, {
       startEvent: startEv,
@@ -1322,8 +1293,6 @@ var Detection = Hammer.detection = {
       deltaY    : delta_y,
 
       distance  : Utils.getDistance(startEv.center, ev.center),
-      angle     : Utils.getAngle(startEv.center, ev.center),
-      direction : Utils.getDirection(startEv.center, ev.center),
 
       scale     : Utils.getScale(startEv.touches, ev.touches),
       rotation  : Utils.getRotation(startEv.touches, ev.touches)